pyproject.toml
[project]
name = "silpi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django-cleanup>=9.0.0",
    "django>=5.1.3",
    "django-environ>=0.11.2",
    "djangorestframework>=3.15.2",
    "pillow>=11.0.0",
    "django-ckeditor>=6.7.1",
    "django-cotton>=1.3.0",
    "django-allauth>=65.2.0",
    "django-unfold==0.41.0",
]
